+++ src/components/CollectionDropdown/CollectionDropdown.jsx
@@ -5,6 +5,7 @@ import Grid from '@mui/material/Grid'
 import NativeSelect from '@mui/material/NativeSelect'
 
 import './CollectionDropdown.css'
+import configEnv from '/public/config.json'
 
 // most of this component comes from the material core UI started code
 // https://mui.com/material-ui/react-select/#native-select
@@ -20,8 +21,8 @@ import {
 } from '../../redux/slices/mainSlice'
 
 const Dropdown = ({ error }) => {
-  const API_ENDPOINT = import.meta.env.VITE_STAC_API_URL
-  const DEFAULT_COLLECTION = import.meta.env.VITE_DEFAULT_COLLECTION
+  const API_ENDPOINT = configEnv.VITE_STAC_API_URL
+  const DEFAULT_COLLECTION = configEnv.VITE_DEFAULT_COLLECTION
 
   const dispatch = useDispatch()
   const [collectionId, setCollectionId] = useState()