+++ src/components/LaunchModal/LaunchModal.jsx
@@ -3,6 +3,7 @@ import './LaunchModal.css'
 import iconCopy from '../../assets/icon-copy.svg'
 import iconExternalLink from '../../assets/icon-external-link.svg'
 import { APP_NAME } from '../defaults'
+import configEnv from '/public/config.json'
 
 import { useDispatch } from 'react-redux'
 import {
@@ -14,7 +15,7 @@ const LaunchModal = () => {
   const dispatch = useDispatch()
   const [copyButtonText, setCopyButtonText] = useState('Copy URL')
   const [copyButtonState, setCopyButtonState] = useState('default')
-  const templateURL = import.meta.env.VITE_CF_TEMPLATE_URL
+  const templateURL = configEnv.VITE_CF_TEMPLATE_URL
 
   function onCloseClick() {
     dispatch(setShowLaunchModal(false))