+++ src/components/Layout/Content/BottomContent/BottomContent.jsx
@@ -7,6 +7,7 @@ import PopupResults from '../../../PopupResults/PopupResults'
 
 import LoadingAnimation from '../../../LoadingAnimation/LoadingAnimation'
 import Legend from '../../../Legend/Legend'
+import configEnv from '/public/config.json'
 
 // redux imports
 import { useSelector, useDispatch } from 'react-redux'
@@ -34,9 +35,9 @@ const BottomContent = () => {
   // if you are setting redux state, call dispatch
   const dispatch = useDispatch()
 
-  const ANALYZE_LINK = import.meta.env.VITE_ANALYZE_BTN_URL
-  const SHOW_PUBLISH_BTN = import.meta.env.VITE_SHOW_PUBLISH_BTN
-  const CF_TEMPLATE_URL = import.meta.env.VITE_CF_TEMPLATE_URL
+  const ANALYZE_LINK = configEnv.VITE_ANALYZE_BTN_URL
+  const SHOW_PUBLISH_BTN = configEnv.VITE_SHOW_PUBLISH_BTN
+  const CF_TEMPLATE_URL = configEnv.VITE_CF_TEMPLATE_URL
   const VIEWER_BTN_TEXT = `Launch Your Own ${APP_NAME}`
 
   const resultType =