+++ src/components/Layout/PageHeader/PageHeader.jsx
@@ -3,11 +3,12 @@ import './PageHeader.css'
 
 import { OpenInNew } from '@mui/icons-material'
 import logoFilmDrop from '../../../assets/logo-filmdrop-e84.png'
+import configEnv from '/public/config.json'
 
 const PageHeader = () => {
-  const DASHBOARD_LINK = import.meta.env.VITE_DASHBOARD_BTN_URL
-  const LOGO = import.meta.env.VITE_LOGO_URL
-  const ALT_TEXT = import.meta.env.VITE_LOGO_ALT
+  const DASHBOARD_LINK = configEnv.VITE_DASHBOARD_BTN_URL
+  const LOGO = configEnv.VITE_LOGO_URL
+  const ALT_TEXT = configEnv.VITE_LOGO_ALT
 
   function onDashboardClick() {
     window.open(DASHBOARD_LINK, '_blank')
@@ -20,7 +21,7 @@ const PageHeader = () => {
           <img src={LOGO} alt={ALT_TEXT} className="headerLogoImage"></img>
         ) : (
           <img
-            src={import.meta.env.PUBLIC_URL + '/logo.png'}
+            src={configEnv.PUBLIC_URL + '/logo.png'}
             alt="FilmDrop by Element 84"
             className="headerLogoImage"
           />