+++ src/components/Search/Search.jsx
@@ -45,6 +45,7 @@ import DateTimeRangePicker from '@wojtekmaj/react-datetimerange-picker'
 import CloudSlider from '../CloudSlider/CloudSlider'
 import CollectionDropdown from '../CollectionDropdown/CollectionDropdown'
 import ViewSelector from '../ViewSelector/ViewSelector'
+import configEnv from '/public/config.json'
 
 const Search = () => {
   const dispatch = useDispatch()
@@ -651,7 +652,7 @@ const Search = () => {
     const mosaicBounds = setupBounds(bbox)
 
     const createMosaicBody = {
-      stac_api_root: import.meta.env.VITE_STAC_API_URL,
+      stac_api_root: configEnv.VITE_STAC_API_URL,
       asset_name: constructMosaicAssetVal(selectedCollectionRef.current),
       collections: [selectedCollectionRef.current],
       datetime,